Moved SettingsDropdown to separate file
